backend/neon_vector_store.py
"""
NeonDB Vector Store with pgvector for production deployment.

Replaces ChromaDB with PostgreSQL + pgvector for cloud-native vector search.
"""

import os
import json
import time
from typing import List, Dict, Any, Optional
from dataclasses import dataclass
import psycopg2
from psycopg2.extras import execute_values, RealDictCursor
from openai import OpenAI


@dataclass
class VectorChunk:
    """Represents a document chunk with embedding."""
    id: str
    text: str
    embedding: List[float]
    metadata: Dict[str, Any]


@dataclass 
class SearchResult:
    """Search result with similarity score."""
    id: str
    text: str
    score: float
    metadata: Dict[str, Any]


class NeonVectorStore:
    """
    Vector store using NeonDB with pgvector extension.
    
    Features:
    - PostgreSQL-native vector similarity search
    - HNSW indexing for fast approximate search
    - Full metadata filtering with SQL
    - Cloud-native, serverless scaling
    """
    
    def __init__(
        self,
        connection_string: Optional[str] = None,
        table_name: str = "paper_chunks",
        embedding_dim: int = 4096  # Qwen3-Embedding-8B dimension
    ):
        """
        Initialize NeonDB vector store.
        
        Args:
            connection_string: NeonDB connection string (or use NEON_DATABASE_URL env)
            table_name: Name of the vector table
            embedding_dim: Dimension of embeddings
        """
        self.connection_string = connection_string or os.getenv("NEON_DATABASE_URL")
        if not self.connection_string:
            raise ValueError("NEON_DATABASE_URL environment variable required")
        
        self.table_name = table_name
        self.embedding_dim = embedding_dim
        self._init_db()
    
    def _get_connection(self):
        """Get database connection."""
        return psycopg2.connect(self.connection_string)
    
    def _init_db(self):
        """Initialize database schema with pgvector."""
        conn = self._get_connection()
        try:
            with conn.cursor() as cur:
                # Enable pgvector extension
                cur.execute("CREATE EXTENSION IF NOT EXISTS vector")
                
                # Create chunks table with vector column
                cur.execute(f"""
                    CREATE TABLE IF NOT EXISTS {self.table_name} (
                        id TEXT PRIMARY KEY,
                        text TEXT NOT NULL,
                        embedding vector({self.embedding_dim}),
                        pmcid TEXT,
                        pmid TEXT,
                        title TEXT,
                        year INTEGER,
                        proteins_mentioned JSONB DEFAULT '[]',
                        aging_theories JSONB DEFAULT '[]',
                        chunk_index INTEGER,
                        created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
                    )
                """)
                
                # Note: IVFFlat index will be created after data is loaded
                # (requires data for training). Use create_index() method after bulk insert.
                
                # Create indexes for filtering
                cur.execute(f"""
                    CREATE INDEX IF NOT EXISTS {self.table_name}_pmcid_idx 
                    ON {self.table_name} (pmcid)
                """)
                cur.execute(f"""
                    CREATE INDEX IF NOT EXISTS {self.table_name}_year_idx 
                    ON {self.table_name} (year)
                """)
                
                conn.commit()
                print(f"[NeonVectorStore] Initialized table: {self.table_name}")
        finally:
            conn.close()
    
    def add_chunks(self, chunks: List[VectorChunk], batch_size: int = 100) -> int:
        """
        Add chunks to the vector store.
        
        Args:
            chunks: List of VectorChunk objects
            batch_size: Number of chunks per batch insert
        
        Returns:
            Number of chunks added
        """
        conn = self._get_connection()
        added = 0
        
        try:
            with conn.cursor() as cur:
                for i in range(0, len(chunks), batch_size):
                    batch = chunks[i:i + batch_size]
                    
                    values = []
                    for chunk in batch:
                        values.append((
                            chunk.id,
                            chunk.text,
                            chunk.embedding,
                            chunk.metadata.get("pmcid", ""),
                            chunk.metadata.get("pmid", ""),
                            chunk.metadata.get("title", ""),
                            chunk.metadata.get("year", 0),
                            json.dumps(chunk.metadata.get("proteins_mentioned", [])),
                            json.dumps(chunk.metadata.get("aging_theories", [])),
                            chunk.metadata.get("chunk_index", 0)
                        ))
                    
                    execute_values(
                        cur,
                        f"""
                        INSERT INTO {self.table_name} 
                        (id, text, embedding, pmcid, pmid, title, year, 
                         proteins_mentioned, aging_theories, chunk_index)
                        VALUES %s
                        ON CONFLICT (id) DO UPDATE SET
                            text = EXCLUDED.text,
                            embedding = EXCLUDED.embedding,
                            pmcid = EXCLUDED.pmcid,
                            pmid = EXCLUDED.pmid,
                            title = EXCLUDED.title,
                            year = EXCLUDED.year,
                            proteins_mentioned = EXCLUDED.proteins_mentioned,
                            aging_theories = EXCLUDED.aging_theories,
                            chunk_index = EXCLUDED.chunk_index
                        """,
                        values,
                        template="(%s, %s, %s::vector, %s, %s, %s, %s, %s::jsonb, %s::jsonb, %s)"
                    )
                    
                    added += len(batch)
                    print(f"[NeonVectorStore] Added batch: {added}/{len(chunks)}")
                
                conn.commit()
        finally:
            conn.close()
        
        return added
    
    def search(
        self,
        query_embedding: List[float],
        top_k: int = 10,
        protein_filter: Optional[str] = None,
        theory_filter: Optional[str] = None,
        year_min: Optional[int] = None,
        year_max: Optional[int] = None
    ) -> List[SearchResult]:
        """
        Search for similar chunks using cosine similarity.
        
        Args:
            query_embedding: Query vector
            top_k: Number of results to return
            protein_filter: Filter by protein symbol
            theory_filter: Filter by aging theory
            year_min: Minimum publication year
            year_max: Maximum publication year
        
        Returns:
            List of SearchResult objects
        """
        conn = self._get_connection()
        
        try:
            with conn.cursor(cursor_factory=RealDictCursor) as cur:
                # Simple query without filters for now
                # Cosine similarity search (1 - cosine_distance)
                query = f"""
                    SELECT 
                        id, text, pmcid, pmid, title, year,
                        proteins_mentioned, aging_theories, chunk_index,
                        1 - (embedding <=> %s::vector) as score
                    FROM {self.table_name}
                    ORDER BY embedding <=> %s::vector
                    LIMIT %s
                """
                
                cur.execute(query, (query_embedding, query_embedding, top_k))
                
                results = []
                for row in cur.fetchall():
                    results.append(SearchResult(
                        id=row["id"],
                        text=row["text"],
                        score=float(row["score"]),
                        metadata={
                            "pmcid": row["pmcid"],
                            "pmid": row["pmid"],
                            "title": row["title"],
                            "year": row["year"],
                            "proteins_mentioned": row["proteins_mentioned"],
                            "aging_theories": row["aging_theories"],
                            "chunk_index": row["chunk_index"]
                        }
                    ))
                
                return results
        finally:
            conn.close()
    
    def count(self) -> int:
        """Get total number of chunks in store."""
        conn = self._get_connection()
        try:
            with conn.cursor() as cur:
                cur.execute(f"SELECT COUNT(*) FROM {self.table_name}")
                return cur.fetchone()[0]
        finally:
            conn.close()
    
    def delete_all(self):
        """Delete all chunks from store."""
        conn = self._get_connection()
        try:
            with conn.cursor() as cur:
                cur.execute(f"TRUNCATE TABLE {self.table_name}")
                conn.commit()
                print(f"[NeonVectorStore] Cleared table: {self.table_name}")
        finally:
            conn.close()
    
    def get_stats(self) -> Dict[str, Any]:
        """Get statistics about the vector store."""
        conn = self._get_connection()
        try:
            with conn.cursor(cursor_factory=RealDictCursor) as cur:
                cur.execute(f"""
                    SELECT 
                        COUNT(*) as total_chunks,
                        COUNT(DISTINCT pmcid) as unique_papers,
                        MIN(year) as min_year,
                        MAX(year) as max_year,
                        AVG(LENGTH(text)) as avg_chunk_length
                    FROM {self.table_name}
                """)
                stats = dict(cur.fetchone())
                
                # Get top proteins
                cur.execute(f"""
                    SELECT protein, COUNT(*) as count
                    FROM {self.table_name},
                    jsonb_array_elements_text(proteins_mentioned) as protein
                    GROUP BY protein
                    ORDER BY count DESC
                    LIMIT 10
                """)
                stats["top_proteins"] = [
                    {"protein": row["protein"], "count": row["count"]}
                    for row in cur.fetchall()
                ]
                
                return stats
        finally:
            conn.close()
    
    def create_index(self):
        """
        Note: pgvector HNSW/IVFFlat indexes are limited to 2000 dimensions.
        Our embeddings are 4096 dimensions, so we use sequential scan.
        For 6621 vectors, this is still fast enough (~100-200ms per query).
        
        If you need faster queries, consider:
        1. Using a smaller embedding model (e.g., 1536 dims)
        2. Using dimensionality reduction (PCA)
        3. Using a dedicated vector DB like Pinecone/Weaviate
        """
        print(f"[NeonVectorStore] Note: Using sequential scan (4096 dims > 2000 limit for indexes)")
        print(f"[NeonVectorStore] For {self.count()} vectors, queries will take ~100-200ms")


if __name__ == "__main__":
    print("NeonVectorStore module - use NEON_DATABASE_URL to connect")
