backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
pydantic-settings

llama-index
llama-index-embeddings-openai
llama-index-embeddings-nebius
llama-index-vector-stores-chroma
chromadb

# NeonDB (PostgreSQL + pgvector) for production
psycopg2-binary
pgvector

numpy
